Added some functions
We added a test for deleting the first item in the cart (DeleteCartTest). I also added a test that will log out by itself (When the test is started, the login is done first and then the logout is done).I made some changes to the Sign Up Section (in the TestData file), I created a function that independently generates a random username. That's why there is no longer a need to manually change the username data in TestData.